+++ .dart_tool/package_config_subset
@@ -66,6 +66,10 @@ fixnum
 2.19
 file:///C:/Users/dimas.hartanto/AppData/Local/Pub/Cache/hosted/pub.dev/fixnum-1.1.0/
 file:///C:/Users/dimas.hartanto/AppData/Local/Pub/Cache/hosted/pub.dev/fixnum-1.1.0/lib/
+floating_bottom_navigation_bar
+2.12
+file:///C:/Users/dimas.hartanto/AppData/Local/Pub/Cache/hosted/pub.dev/floating_bottom_navigation_bar-1.5.2/
+file:///C:/Users/dimas.hartanto/AppData/Local/Pub/Cache/hosted/pub.dev/floating_bottom_navigation_bar-1.5.2/lib/
 flutter_cache_manager
 2.17
 file:///C:/Users/dimas.hartanto/AppData/Local/Pub/Cache/hosted/pub.dev/flutter_cache_manager-3.3.1/
