+++ lib/components/menu/menu.dart
@@ -1,11 +1,5 @@
-import 'dart:ui';
+// ignore_for_file: unused_import
 
-import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
-import 'package:flutter_map/flutter_map.dart';
-import 'package:google_fonts/google_fonts.dart';
-import 'package:latlong2/latlong.dart';
-import 'package:sliding_up_panel/sliding_up_panel.dart';
 
-part 'more_menu.dart';
+part 'menu_pesawat.dart';
