+++ lib/components/widget/ide_liburan.dart
@@ -1,0 +1,47 @@
+part of 'widget.dart';
+
+class IdeLiburan extends StatelessWidget {
+  const IdeLiburan({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return SingleChildScrollView(
+      child: Padding(
+        padding: const EdgeInsets.only(
+          left: 16,
+          right: 16,
+        ),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            Expanded(
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    'Ide liburan yang wajib dicatat',
+                    style: GoogleFonts.montserrat(
+                      color: Colors.black,
+                      fontSize: 16,
+                      fontWeight: FontWeight.bold,
+                    ),
+                  ),
+                  SizedBox(height: 4),
+                  Text(
+                    // overflow: TextOverflow.ellipsis,
+                    'inspirasi liburan serta rekomendasi tiket, penginapan, transportasi dan info lain untukmu',
+                    style: GoogleFonts.montserrat(
+                      color: Colors.grey[800],
+                      fontSize: 12,
+                      fontWeight: FontWeight.normal,
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
