+++ lib/components/widget/judul_travel.dart
@@ -1,0 +1,42 @@
+part of 'widget.dart';
+
+class JudulTravel extends StatelessWidget {
+  const JudulTravel({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(
+        left: 16,
+        right: 16,
+      ),
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text(
+                'Citilink Online Travel Fair',
+                style: GoogleFonts.montserrat(
+                  color: Colors.black,
+                  fontSize: 16,
+                  fontWeight: FontWeight.bold,
+                ),
+              ),
+              SizedBox(height: 4),
+              Text(
+                'Serbu hargahematnya, diskon hingga IDR 1,5 Juta !',
+                style: GoogleFonts.montserrat(
+                  color: Colors.grey[800],
+                  fontSize: 12,
+                  fontWeight: FontWeight.normal,
+                ),
+              ),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+}
