+++ lib/components/widget/main_menu.dart
@@ -1,0 +1,170 @@
+// ignore_for_file: prefer_const_constructors, prefer_const_literals_to_create_immutables
+
+part of 'widget.dart';
+
+class MainMenu extends StatelessWidget {
+  const MainMenu({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: EdgeInsets.all(15),
+      child: Column(
+        children: <Widget>[
+          Row(
+            mainAxisAlignment: MainAxisAlignment.spaceAround,
+            // ignore: prefer_const_literals_to_create_immutables
+            children: <Widget>[
+              TravelinIconMenu(
+                image:
+                    "https://cdn-icons-png.flaticon.com/128/3127/3127363.png",
+                title: "Pesawat",
+              ),
+              TravelinIconMenu(
+                image:
+                    "https://cdn-icons-png.flaticon.com/128/3378/3378741.png",
+                title: "Hotel",
+              ),
+              TravelinIconMenu(
+                image:
+                    "https://cdn-icons-png.flaticon.com/128/8315/8315136.png",
+                title: "Event",
+              ),
+              TravelinIconMenu(
+                image:
+                    "https://cdn-icons-png.flaticon.com/128/2169/2169416.png",
+                title: "Atraksi",
+              ),
+            ],
+          ),
+          SizedBox(height: 15),
+          Row(
+            mainAxisAlignment: MainAxisAlignment.spaceAround,
+            children: <Widget>[
+              TravelinIconMenu(
+                image:
+                    "https://cdn-icons-png.flaticon.com/128/1314/1314576.png",
+                title: "JR Pass",
+              ),
+              TravelinIconMenu(
+                image:
+                    "https://cdn-icons-png.flaticon.com/128/1841/1841602.png",
+                title: "Bus",
+              ),
+              TravelinIconMenu(
+                image: "https://cdn-icons-png.flaticon.com/128/620/620765.png",
+                title: "Travel",
+              ),
+              GestureDetector(
+                onTap: () {
+                  showModalBottomSheet(
+                      context: context,
+                      builder: (BuildContext context) {
+                        return Container(
+                          child: Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              GestureDetector(
+                                child: Center(
+                                  child: Container(
+                                      margin: EdgeInsets.only(top: 15),
+                                      height: 20,
+                                      width: 60,
+                                      decoration: BoxDecoration(
+                                          borderRadius:
+                                              BorderRadius.circular(10),
+                                          color: Colors.grey.shade300)),
+                                ),
+                              ),
+                              Container(
+                                padding: EdgeInsets.all(25),
+                                child: Text(
+                                  'Layanan Teratas',
+                                  style: GoogleFonts.poppins(
+                                    color: Colors.black,
+                                    fontSize: 14,
+                                    fontWeight: FontWeight.bold,
+                                  ),
+                                ),
+                              ),
+                              Row(
+                                mainAxisAlignment:
+                                    MainAxisAlignment.spaceAround,
+                                children: <Widget>[
+                                  TravelinIconMenu(
+                                    image:
+                                        "https://cdn-icons-png.flaticon.com/128/3127/3127363.png",
+                                    title: "Pesawat",
+                                  ),
+                                  TravelinIconMenu(
+                                    image:
+                                        "https://cdn-icons-png.flaticon.com/128/3378/3378741.png",
+                                    title: "Hotel",
+                                  ),
+                                  TravelinIconMenu(
+                                    image:
+                                        "https://cdn-icons-png.flaticon.com/128/8315/8315136.png",
+                                    title: "Event",
+                                  ),
+                                  TravelinIconMenu(
+                                    image:
+                                        "https://cdn-icons-png.flaticon.com/128/2169/2169416.png",
+                                    title: "Atraksi",
+                                  ),
+                                ],
+                              ),
+                              Container(
+                                padding: EdgeInsets.all(25),
+                                child: Text(
+                                  'Layanan Lainnya',
+                                  style: GoogleFonts.poppins(
+                                    color: Colors.black,
+                                    fontSize: 14,
+                                    fontWeight: FontWeight.bold,
+                                  ),
+                                ),
+                              ),
+                              Row(
+                                mainAxisAlignment:
+                                    MainAxisAlignment.spaceAround,
+                                children: <Widget>[
+                                  TravelinIconMenu(
+                                    image:
+                                        "https://cdn-icons-png.flaticon.com/128/3127/3127363.png",
+                                    title: "Pesawat",
+                                  ),
+                                  TravelinIconMenu(
+                                    image:
+                                        "https://cdn-icons-png.flaticon.com/128/3378/3378741.png",
+                                    title: "Hotel",
+                                  ),
+                                  TravelinIconMenu(
+                                    image:
+                                        "https://cdn-icons-png.flaticon.com/128/8315/8315136.png",
+                                    title: "Event",
+                                  ),
+                                  TravelinIconMenu(
+                                    image:
+                                        "https://cdn-icons-png.flaticon.com/128/2169/2169416.png",
+                                    title: "Atraksi",
+                                  ),
+                                ],
+                              ),
+                            ],
+                          ),
+                        );
+                      });
+                },
+                child: TravelinIconMenu(
+                  image:
+                      "https://cdn-icons-png.flaticon.com/128/11781/11781149.png",
+                  title: "More",
+                ),
+              ),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+}
