+++ lib/components/widget/widget.dart
@@ -10,3 +10,10 @@ part 'carddestinasi.dart';
 part 'button_home.dart';
 part 'slide_card_perjalanan.dart';
 part 'rencana_perjalanan.dart';
+part 'ide_liburan.dart';
+part 'judul_travel.dart';
+part 'card_categori.dart';
+part 'konten_travel.dart';
+part 'main_menu.dart';
+part 'traveling_kuy_card.dart';
+part 'card_ide_liburan.dart';
