+++ pubspec.yaml
@@ -33,6 +33,7 @@ dependencies:
   carousel_slider: ^4.0.0
   cupertino_icons: ^1.0.2
   custom_navigation_bar: ^0.8.2
+  floating_bottom_navigation_bar: ^1.5.2
   flutter:
     sdk: flutter
   flutter_map: ^6.1.0
